Add About-Release to complement About-Dev
The system going forward will be to make all edits to About-Release.xml,
and to ignore About-Dev.xml and the temporarily generated About.xml.
This way edits to About don't get overwritten accidentally. If edits are
made to About.xml on accident, they will be preserved as long as
About-Release and About-Dev are not edited after that point.